+++ src/class-type.ts
@@ -1,7 +1,54 @@
+function typeHandle(type: StructType): { set: string; get: string } {
+  switch (type.size) {
+    case 1:
+      return {
+        get: type.unsigned ? "getUint8" : "getInt8",
+        set: type.unsigned ? "setUint8" : "setInt8",
+      };
+    case 2:
+      return {
+        get: type.unsigned ? "getUint16" : "getInt16",
+        set: type.unsigned ? "setUint16" : "setInt16",
+      };
+
+    case 4:
+      const isFloat = type.typeName === "float";
+      return {
+        get: isFloat ? "getFloat32" : type.unsigned ? "getUint32" : "getInt32",
+        set: isFloat ? "setFloat32" : type.unsigned ? "setUint32" : "setInt32",
+      };
+
+    case 8:
+      const isDouble = type.typeName === "double";
+      return {
+        get: isDouble
+          ? "getFloat64"
+          : type.unsigned
+          ? "getBigUint64"
+          : "getBigInt64",
+        set: isDouble
+          ? "setFloat64"
+          : type.unsigned
+          ? "setBigUint64"
+          : "setBigInt64",
+      };
+
+    default:
+      throw new Error(`StructBuffer: Unrecognized ${type} type.`);
+  }
+}
+
 class StructTypeNext {
   deeps: number[] = [];
 
-  constructor(k: number, public readonly typeName: string) {
+  constructor(
+    k: number,
+    public readonly typeName: string,
+    public readonly size: number,
+    public readonly unsigned: boolean,
+    public readonly get: string,
+    public readonly set: string
+  ) {
     this.deeps.push(k);
     const proxy: this = new Proxy(this, {
       get(o: any, k: string | number | symbol) {
@@ -18,15 +65,13 @@ class StructTypeNext {
   }
 }
 
-export abstract class StructType extends Array {
-  abstract typeName: string;
-
+export class StructType extends Array {
   deeps: number[] = [];
 
   /**
    * ```
-   * <type>[<n>]
-   * <type>[<n>][<n>]...
+   * float[2] => true
+   * float    => false
    * ```
    */
   get isList(): boolean {
@@ -38,8 +83,22 @@ export abstract class StructType extends Array {
     return this.deeps.reduce((acc, it) => (acc *= it), 1);
   }
 
-  constructor() {
+  is(type: StructType): boolean {
+    return type.typeName === this.typeName;
+  }
+
+  get: string;
+  set: string;
+
+  constructor(
+    public readonly typeName: string,
+    public readonly size: 1 | 2 | 4 | 8,
+    public readonly unsigned: boolean
+  ) {
     super();
+    const { set, get } = typeHandle(this);
+    this.set = set;
+    this.get = get;
     const proxy: this = new Proxy(this, {
       get(o: any, k: string | number | symbol) {
         // 普通属性直接返回
@@ -48,7 +107,14 @@ export abstract class StructType extends Array {
         // 如果访问char[2] ，数字属性，那么立即返回一个新的proxy，避免上下文冲突
         k = k.toString();
         if (/\d+/.test(k)) {
-          const newProxy: any = new StructTypeNext(parseInt(k), o.typeName);
+          const newProxy: any = new StructTypeNext(
+            parseInt(k),
+            o.typeName,
+            o.size,
+            o.unsigned,
+            o.get,
+            o.set
+          );
 
           // 避免 instanceof 检测
           // 同时prototype上的属性和方法也会拷贝过去
@@ -65,51 +131,10 @@ export abstract class StructType extends Array {
   }
 }
 
-export class Char extends StructType {
-  typeName: string = "char";
-}
-export class String_t extends StructType {
-  typeName: string = "string_t";
-}
-export class Float extends StructType {
-  typeName: string = "float";
-}
-export class Double extends StructType {
-  typeName: string = "double";
-}
-export class Int8_t extends StructType {
-  typeName: string = "int8_t";
-}
-export class Int16_t extends StructType {
-  typeName: string = "int16_t";
-}
-export class Int32_t extends StructType {
-  typeName: string = "int32_t";
-}
-export class Int64_t extends StructType {
-  typeName: string = "int64_t";
-}
-export class Byte extends StructType {
-  typeName: string = "byte";
-}
-export class Word extends StructType {
-  typeName: string = "word";
-}
-export class Dword extends StructType {
-  typeName: string = "dword";
-}
-export class Qword extends StructType {
-  typeName: string = "qword";
-}
-export class Uint8_t extends StructType {
-  typeName: string = "uint8_t";
-}
-export class Uint16_t extends StructType {
-  typeName: string = "uint16_t";
-}
-export class Uint32_t extends StructType {
-  typeName: string = "uint32_t";
-}
-export class Uint64_t extends StructType {
-  typeName: string = "uint64_t";
+export function registerType(
+  typeName: string,
+  size: 1 | 2 | 4 | 8,
+  unsigned = true
+): StructType {
+  return new StructType(typeName, size, unsigned);
 }