+++ src/types.ts
@@ -1,60 +1,24 @@
-import {
-  Char,
-  Double,
-  Float,
-  String_t,
-  Int8_t,
-  Int16_t,
-  Int32_t,
-  Int64_t,
-  Byte,
-  Word,
-  Dword,
-  Qword,
-  Uint8_t,
-  Uint16_t,
-  Uint32_t,
-  Uint64_t,
-} from "./class-type";
+import { registerType } from "./class-type";
 
-export const char = new Char();
-export const string_t = new String_t();
+export const char = registerType("char", 1);
+export const string_t = registerType("string_t", 1);
 
 // Floating point / 浮点
-export const float = new Float();
-export const double = new Double();
+export const float = registerType("float", 4);
+export const double = registerType("double", 8);
 
 // signed integer / 有符号整数
-export const int8_t = new Int8_t();
-export const int16_t = new Int16_t();
-export const int32_t = new Int32_t();
-export const int64_t = new Int64_t();
+export const int8_t = registerType("int8_t", 1, false);
+export const int16_t = registerType("int16_t", 2, false);
+export const int32_t = registerType("int32_t", 4, false);
+export const int64_t = registerType("int64_t", 8, false);
 
 // unsigned integer / 无符号整数
-export const BYTE = new Byte();
-export const WORD = new Word();
-export const DWORD = new Dword();
-export const QWORD = new Qword();
-export const uint8_t = new Uint8_t();
-export const uint16_t = new Uint16_t();
-export const uint32_t = new Uint32_t();
-export const uint64_t = new Uint64_t();
-
-export const ALL_TYPE_NAMES: string[] = [
-  int8_t.typeName,
-  int16_t.typeName,
-  int32_t.typeName,
-  int64_t.typeName,
-  uint8_t.typeName,
-  uint16_t.typeName,
-  uint32_t.typeName,
-  uint64_t.typeName,
-  BYTE.typeName,
-  WORD.typeName,
-  DWORD.typeName,
-  QWORD.typeName,
-  float.typeName,
-  double.typeName,
-  char.typeName,
-  string_t.typeName,
-];
+export const BYTE = registerType("BYTE", 1);
+export const WORD = registerType("WROD", 2);
+export const DWORD = registerType("DWORD", 4);
+export const QWORD = registerType("QWORD", 8);
+export const uint8_t = registerType("uint8_t", 1);
+export const uint16_t = registerType("uint16_t", 2);
+export const uint32_t = registerType("uint32_t", 4);
+export const uint64_t = registerType("uint64_t", 8);